+++ packages/ses/src/immutable-array-buffer-shim.js
@@ -1,0 +1,4 @@
+import { globalThis } from './commons.js';
+import { ImmutableArrayBuffer } from './immutable-array-buffer.js';
+
+globalThis.ImmutableArrayBuffer = ImmutableArrayBuffer;